* prihlaska pro více ORG

+++ prihlaska_2025.4.php
@@ -107,10 +107,10 @@ try {
     $ORG= $access2org[$_GET['org']];
     $ORG->code= $_GET['org'];
     // detekce varianty: normální nebo testovací - buďto přihlášení do Answer nebo volání z webu
-    $ANSWER= $SID ? 1 : ($_SESSION[
-        $_TEST ? ($domain=='asc' ? 'asc' : 'dbt') 
-               : ($domain=='asc' ? 'asc' : 'db2') 
-        ]['user_id']??0);
+    $ANSWER= $SID ? 1 : ($_SESSION[$ORG->sess]['user_id']??0);
+//        $_TEST ? ($domain=='asc' ? 'asc' : 'dbt') 
+//               : ($domain=='asc' ? 'asc' : 'db2') 
+//        ]['user_id']??0);
     // odvození požadavku na test a ostrý mail
     $TEST= $_GET['test']??0 ? ($ANSWER?(0+$_GET['test']):0) : 0;
     $MAIL= $_GET['mail']??1 ? 1 : ($ANSWER?0:1);
@@ -2960,7 +2960,7 @@ function log_write_changes() { // ----------------------------------------------
   elseif ($TRACE)
       display("LOG_WRITE_VARS fail - no sesssion");
 } // zapíše změněné $vars před zobrazením formuláře 
-function log_find_saved($email) { // ------------------------------------------------ log find_saved
+function log_find_saved($email) { trace(); // --------------------------------------- log find_saved
   global $vars;
   // zkusíme najít poslední verzi přihlášky - je ve fázi (c)
   $found= '';
@@ -2974,7 +2974,7 @@ function log_find_saved($email) { // -------------------------------------------
   } // už se povedlo přihlásit
   list($idpr,$open)= select_2("SELECT id_prihlaska,open FROM prihlaska 
       WHERE id_pobyt=0 AND id_akce=$vars->id_akce AND email='$email' AND vars_json!='' 
-      AND id_prihlaska>110 ORDER BY id_prihlaska DESC LIMIT 1");
+      ORDER BY id_prihlaska DESC LIMIT 1");
   if (!$idpr) goto end;
   $vars->continue= $idpr;
   list($vars->ido,$vars->idr)= select_2(
@@ -3069,8 +3069,9 @@ function log_close() { // ------------------------------------------------------
     }
   }
 }
-function append_log($msg) { // ------------------------------------------------------ append error
+function append_log($msg) { // ---------------------------------------------------------- append log
   global $AKCE, $VERZE, $MINOR, $CORR_JS, $TEST, $ezer_version;
+  append_org_log($msg);
   $file= "prihlaska.log.php";
   $idw= $_SESSION[$AKCE]->id_prihlaska??'?';
   $email= $_SESSION[$AKCE]->email??'?';
@@ -3097,6 +3098,20 @@ if(!(\$_SESSION['ast']['user_id']??0) && !(\$_SESSION['db2']['user_id']??0) && !
 ><b>VERZE/JS  AKCE DATUM      ČAS      PŘIHLÁŠKA       KLIENT </b>\n
 __EOS;
       file_put_contents($file, $prefix);
+  }
+  file_put_contents($file, "$msg\n", FILE_APPEND);
+}
+function append_org_log($msg) { // -------------------------------------------------- append org_log
+  global $ORG, $AKCE, $VERZE, $MINOR, $CORR_JS, $TEST;
+  $file= "prihlaska.log.{$ORG->code}.php";
+  $idw= $_SESSION[$AKCE]->id_prihlaska??'?';
+  $email= $_SESSION[$AKCE]->email??'?';
+  $ip= $_SERVER['HTTP_X_REAL_IP'] ?? $_SERVER['REMOTE_ADDR'];
+  $x= $TEST==2 ? " TEST=2 " : "$VERZE.$MINOR/$CORR_JS";
+  $ida= str_pad(substr($AKCE,2),4,' ',STR_PAD_LEFT);
+  $msg= "$x $ida ".date('Y-m-d H:i:s').str_pad($idw,5,' ',STR_PAD_LEFT)." $msg mail=$email ip=$ip";
+  if (!file_exists($file)) {
+    file_put_contents($file, "<?php die('Sorry, no access');?>\n");
   }
   file_put_contents($file, "$msg\n", FILE_APPEND);
 }
